fix: api 변경 대응 및 appbar, NavBar 컴포넌트 재사용 컴포넌트 생성
